Fix expectations
Now the output zarr store does not change it's type based on the offset and
scale.  We never apply that to the store on input.  This commit aligns the
tests to the code.

DAS-1673

+++ tests/test_adapter.py
@@ -98,7 +98,7 @@ class TestAdapter(TestCase):
             /
              ├── data
              │   ├── horizontal
-             │   │   ├── east (1, 3, 3) int64
+             │   │   ├── east (1, 3, 3) uint8
              │   │   └── west (1, 3, 3) uint8
              │   └── vertical
              │       ├── north (1, 3, 3) uint8
@@ -273,7 +273,7 @@ class TestAdapter(TestCase):
             /
              ├── data
              │   ├── horizontal
-             │   │   ├── east (2, 3, 3) int64
+             │   │   ├── east (2, 3, 3) uint8
              │   │   └── west (2, 3, 3) uint8
              │   └── vertical
              │       ├── north (2, 3, 3) uint8
